+++ src/sidebar/sidebar-main.js
@@ -4,6 +4,7 @@ function renderAll() {
   renderHeader();
   renderOverview();
   renderIndexability();
+  renderSerp();
   renderIssues();
   renderHeadings();
   renderLinks();